pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nanobuddy"
version = "0.1.0"
description = "Lightweight wake word detection — streaming inference, training, and tooling."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"

dependencies = [
    "numpy",
    "onnxruntime",
]

[project.optional-dependencies]
detect = ["sounddevice"]
train = [
    "torch>=2.2",
    "torchaudio>=2.2",
    "onnx",
    "scipy",
    "pyyaml",
    "audiomentations",
    "torch-audiomentations",
    "librosa",
    "numba>=0.59",
    "soundfile",
    "tqdm",
    "matplotlib",
    "torchinfo",
    "torchmetrics",
    "psutil",
    "scikit-learn",
]
tools = [
    "nanobuddy[train,detect]",
    "click>=8",
    "pronouncing",
    "faster-whisper",
    "datasets",
    "huggingface-hub",
    "rich",
    "requests",
]
dev = [
    "pytest",
    "pyyaml",
]

[project.urls]
Homepage = "https://github.com/arcosoph/nanowakeword"

[project.scripts]
nanobuddy = "nanobuddy.tools.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["nanobuddy*"]

[tool.setuptools.package-data]
"nanobuddy.models.mel_spectrogram" = ["*.onnx"]
"nanobuddy.models.embedding" = ["*.onnx"]
"nanobuddy.models.vad" = ["*.onnx"]
"nanobuddy.models.wake_word" = ["*.onnx", "*.pt"]
